Added test_clean_adduct
